# Vault/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Vaults Manager",
    "version" : "1.0.0",
    "title" : "Vaults Manager API",
    "contact" : {
      "email" : "aawadall@ualberta.net"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/awadallah/VaultsManager/1.0.0",
  "tags" : [ {
    "name" : "admins",
    "description" : "Secured Admin-only calls"
  }, {
    "name" : "developers",
    "description" : "Operations available to regular developers"
  }, {
    "name" : "devops",
    "description" : "Administrative calls"
  } ],
  "schemes" : [ "https" ],
  "paths" : {
    "/health" : {
      "get" : {
        "tags" : [ "devops" ],
        "summary" : "simple healthcheck",
        "description" : "a simple health check endpoint to ensure the API is running \n",
        "operationId" : "simpleHealthCheck",
        "produces" : [ "application/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "API running",
            "schema" : {
              "type" : "object",
              "properties" : {
                "status" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "description" : "API Not running"
          }
        }
      }
    },
    "/inventory" : {
      "get" : {
        "tags" : [ "developers" ],
        "summary" : "searches inventory",
        "description" : "By passing in the appropriate options, you can search for\navailable inventory in the system\n",
        "operationId" : "searchInventory",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "searchString",
          "in" : "query",
          "description" : "pass an optional search string for looking up inventory",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip for pagination",
          "required" : false,
          "type" : "integer",
          "minimum" : 0,
          "format" : "int32"
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "maximum number of records to return",
          "required" : false,
          "type" : "integer",
          "maximum" : 50,
          "minimum" : 0,
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/InventoryItem"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      },
      "post" : {
        "tags" : [ "admins" ],
        "summary" : "adds an inventory item",
        "description" : "Adds an item to the system",
        "operationId" : "addInventory",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryItem",
          "description" : "Inventory item to add",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/InventoryItem"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "item created"
          },
          "400" : {
            "description" : "invalid input, object invalid"
          },
          "409" : {
            "description" : "an existing item already exists"
          }
        }
      }
    }
  },
  "definitions" : {
    "InventoryItem" : {
      "type" : "object",
      "required" : [ "id", "manufacturer", "name", "releaseDate" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "format" : "uuid",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "name" : {
          "type" : "string",
          "example" : "Widget Adapter"
        },
        "releaseDate" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2016-08-29T09:12:33.001Z"
        },
        "manufacturer" : {
          "$ref" : "#/definitions/Manufacturer"
        }
      },
      "example" : {
        "releaseDate" : "2016-08-29T09:12:33.001Z",
        "name" : "Widget Adapter",
        "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
        "manufacturer" : {
          "phone" : "408-867-5309",
          "name" : "ACME Corporation",
          "homePage" : "https://www.acme-corp.com"
        }
      }
    },
    "Manufacturer" : {
      "required" : [ "name" ],
      "properties" : {
        "name" : {
          "type" : "string",
          "example" : "ACME Corporation"
        },
        "homePage" : {
          "type" : "string",
          "format" : "url",
          "example" : "https://www.acme-corp.com"
        },
        "phone" : {
          "type" : "string",
          "example" : "408-867-5309"
        }
      },
      "example" : {
        "phone" : "408-867-5309",
        "name" : "ACME Corporation",
        "homePage" : "https://www.acme-corp.com"
      }
    },
    "inline_response_200" : {
      "properties" : {
        "status" : {
          "type" : "string"
        }
      },
      "example" : {
        "status" : "status"
      }
    }
  }
}